pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "nonogram-app"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0",
    "polars>=0.20",
    "pillow>=10.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["nonogram*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
